refactor Skin 폴더 정리
